Avoid throwing an exception for sourceInformation like string keys
Avoid throwing when the a string appears where a source information is
stored stored and doesn't parse correclty as a source information
string.

This can occur when `ReusableContent`, used in `rememberSaveable` is
passed a string value that contains leading integer that is outside
the normal integer range.

This exception happens in the inspector and occurs in the tooling
API the inspector uses. It doesn't affect normal operation.

Fixes: 213478154
Test: androidx.compose.ui.tooling.data.InspectableTests
Change-Id: Ia1dee659057c20358f32ce855d5059d6c7b369ee