By default, ignore drag events containing localState, with an option to accept them.
The goal is to prevent confusing 'self-drags'. localState can only be accessed from the same Activity that initiated the drag event. It is a common case that apps might have both draggable data and a drop target, but using them in tandem doesn't really make sense. For example, in a chat app, dragging an image you were sent, but the only available drop target being to send it back to the same chat. The app simply needs to set a non-null localState when initiating the drag.

Bug: 194920587
Test: DropHelperTest
Relnote: Ignore DragEvents with localState by default, with a configuration option to accept them.

Change-Id: I55792b436b346ed9e5fa092c88b4e878087d1395