Add CircularProgressIndicator
Progress indicators display the length of a process or an unspecified wait time.
Wear ProgressIndicator differs from Material compose ProgressIndicator. It has a startAngle and endAngle, by which it can have a gap at the top ( where the TimeText might be) or at any place on the circle. It also allows customization of background color, as it might be required.

Bug: 210397099
Test: "androidx.wear.compose.material.ProgressIndicatorTest, androidx.wear.compose.material.ProgressIndicatorScreenshotTest"
Relnote: "Add CircularProgressIndicator. Progress indicators display the progress of a process or an indication that a long running process with no intermediate indication of its progress is still running."
Change-Id: Iab8daa0e5355391bc09d78690d4690e489f7621c