Add InitializerViewModelFactory
With the addition of CreationExtras getting a ViewModel from an extra
has become very simple. It would be nice to provide a nice high level
API surface for developers to use for simple ViewModel creation cases.

The InitiailizerViewModelFactory links ViewModel class names to simple
lambdas that can create a ViewModel using information provided in the
CreationExtras.

RelNote: "`lifecycle-viewmodel` now provides an
`InitializerViewModelFactory` that allows you to add lambda for handling
particular ViewModel classes, using `CreationExtras` as the primary data
source."
Test: ViewModelInitializerTest
Bug: 216687549

Change-Id: If58fc49053a6b551643e411c648da564680bff32