Introduce onFocusBoundsChanged.
Design doc: https://docs.google.com/document/d/16VnAGMoRspc3n976hKlkxX6f9LnGkc3eYgHDF70_AoQ/edit?usp=sharing&resourcekey=0-arS4YHxpB4QN0RbdsQIehw

Note that this feature does not currently work with AndroidView (b/220030968), since
that needs to be wired up in the ui module but this API is needed to fix
an issue in foundation, and I want to leave it as Experimental for now,
and we don't allow cross-module experimental access. When this modifier
is stabilized, it should be moved to UI, and that bug resolved.

Bug: b/190539358
Bug: b/192043120
Bug: b/216842427
Test: Manual via FocusedChildDemo
Test: ./gradlew :compose:f:f:cDAT
Relnote: "Introduced experimental `Modifier.onFocusedBoundsChanged` to allow
    observing the bounds of child focusables."
Change-Id: I14283393b5273527ab65f4aa1a2d4383321b0d95