Allow listening for ComponentActivity's onPictureInPictureModeChanged()
Rather than requiring subclasses of ComponentActivity
override onPictureInPictureModeChanged() (both the one that only
takes a boolean for API 24-25 and the one that takes a boolean +
Configuration object for API 26+), allow any component to add
a listener for these events.

The class returned always includes the boolean isInPictureInPictureMode,
but has an API level guarded call to get the new Configuration
to conform to when that information became available.

It also means that FragmentActivity can remove its custom
override and instead rely on this new API entirely.

Relnote: "`ComponentActivity` now implements the
`OnPictureInPictureModeChangedProvider` interface to allow any
component to receive picture-in-picture mode change events."
BUG: 71584872
Test: updated and new test suites pass

Change-Id: I9f56767522d22873e0539a7f1a51257972619806