Change the default initialCenterItemIndex of the ScalingLazyListState from 0->1.
Change the default list item that we place in the center of the ScalingLazyColumn viewport from index 0->1. This gives a more natural default set up for most lists, especially if they have a list header item as the first item.

RelNote: "We have changed the default initialCenterItemIndex for the ScalingLazyListState from 0->1. This means that unless overridden on state construction with ScalingLazyListState.rememberScalingLazyListState(initialCenterItemIndex = <value>) then the 2nd list item (index == 1) will be placed in the center of the viewport at initialization and the 1st (index == 0) item will be placed before it. This allows for a better default visual effect out of the box as most of viewport will be filled with list items."

Bug: 217344252
Test: ./gradlew wear:compose:compose-material:connectedCheck --info --daemon
Change-Id: I0c62304ff98c3233e48f38a3641ef856491b5838