Add optional snap support to ScalingLazyColumn.
Move the Picker snap support to ScalingLazyColumnDefaults. Extend it to allow an offset to be applied when snapping. This is used if the developer wants to use Start Anchoring and make 2 list items "Straddle" the Viewport center line.

Bug: 217377069
Test: ./gradlew wear:compose:compose-material:connectedCheck --info --daemon
RelNote: "We have added snap support that can be used with ScalingLazyColumn. Set flingBehavior = ScalingLazyColumnDefaults.flingWithSnapBehavior(state) to enable snap support."
Change-Id: I068d3131c38b413df242977c708698bb32047394