Add a work around in FileProvider
Some OEMs strip meta-data from the manifest. This is problematic as
FileProvider was depending on meta-data to specify a resource containing
the paths that should be shared. This change allows FileProvider to be
extended in order to specify the resource id in code. Documentation was
also updated.

Bug: b/117428411, b/30826848, b/62344061, b/65457903
Relnote: Added a work-around to FileProvider for OEMs stripping meta-data from manifests.
Test: FileProviderTest
Change-Id: I82f63ccc3d95c4b82c5c42db8cad4d09cd152017