CameraX/MLKit integration
With this change, developer will be able to create a MLKit detector(FaceDetector, BarcodeScanner etc.) and use it directly with CameraControlller to receive detection results. This integration also takes care of the transformation for the coordinates in the result.

- Add a new experimental annotation. All the new change in this CL are experimental.
- In camera-core artifact, add new methods to the Analyzer interface. The new interfaces have default implementations so they won't be breaking changes.
- In camera-mlkit-vision artfact, implement Analyzer to wrap around MLKit APIs.
- In camera-view artifact, update CameraController to handle the new Analyzer methods
- Add a new Fragment in camera-view test app.

Relnote: an experimental API for using CameraController with external image processing libraries. e.g. MLKit

Bug: 198984186
Test: manual test and ./gradlew bOS
Change-Id: I4ea71fb8112a13c16dbc79873267b7714e94c6ec