Add DatePicker demo and fix PickerState bug
Implemented a Date Picker demo (we may revisit the UX for this in the future. Also fixes a bug where the initiallySelectedOption on PickerState was not selected until the Picker was displayed. Also fixed cosmetic issue with hours/minutes/seconds not all being visible on smaller devices. Animation has not yet been added.

Test: Run androidx.wear.compose.integration.demo tests

Relnote: "We have added a demo for a Date Picker and fixed a bug in PickerState where the initiallySelectedOption was not applied until the Picker had been displayed."

Bug: 217920875
Bug: 219964689
Bug: 220342661
Change-Id: Id0d7efdace5625690e444432de7bc2416fce551b