ScalingLazyColumn - update autoCentering spacing/padding
Based on UX review and feedback we are updating the amount of Spacing/Padding that is added at the beginning of the ScalingLazyColumn so that is it sufficient to ensure that the list item with the index provided to the ScalingLazyListState.initialCenterItemIndex can be scrolled to the center. This will avoid having excess space at the top of the list that was previously provided to allow "all" list items to be scrolled into view.

RelNote: "Based on UI/UX feedback we have changed the ScalingLazyColumn autoCentering so that it will provide only enough space to make sure the items with index ScalingLazyListState.initialCenterItemIndex or higher will be able to be fully scrolled to the center of the viewport. This allows developers to place one or two items about the item in initially in the center which are not scrollable into the middle. This will mean that an autoCenter'ing ScalingLazyColumn will not be able to scroll above the initialCenterItemIndex/initialCenterItemScrollOffset"
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 220843404
Change-Id: I22ee2c5d2218cb89b3abfcc474386c26a424430f