Revert "Convert WindowInsetsControllerCompat behavior to depend on a View to align behavior"
This reverts commit 6766e78f6a481a17ece134719066d09c981f30b1.

Reason for revert: The approach for constructing the WindowInsetsControllerCompat with just a View doesn't work in Activity.onCreate due to the window flags.

Relnote: "Reverted previous change of relying solely on a View for
WindowInsetsControllerCompat, and again require a Window which is
required for managing some window flags. Deprecated
ViewCompat.getWindowInsetsController in favor of
WindowCompat.getInsetsController to ensure that the correct Window is
used (such as if the View is in a dialog)."

Bug: 219572936
Test: Updated tests

Change-Id: I660aee32108b59516232b41e05b3f05ae2538554