Add more ScalingLazyList Demos to Wear Compose Integration Test Demo
Collection of updates and additions to teh Wear Compose Integration Test Demo at the request of the UX team.
Added ability to add a description to demos that is placed underneath the chip that runs the demo.
Added more ScalingLazyColumn demos including a stress test and ability to see "Guidelines" on some of the tests to make it easier to visually check the list item scaling.
Tidied up the padding and other visuals.
Added a list stress test demo with lots of different sized components in it.

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 220844907
Bug: 220845712
Change-Id: Id8d0b2f414ff661e094d7c92ac4e104b5592d186