Restrict scope of ExperimentalWearMaterialApi to Modifier.swipeable and SwipeableState
Refactored SwipeToDismissBoxState so that it has a SwipeableState property instead of inheriting. This allows us to present a stable API, with use of SwipeableState kept as an implementation detail that can be changed. As a result, added the following to the SwipetoDismissBoxState API to support typical usage: currentValue, targetValue and isAnimationRunning properties and snapTo function. Also copied over fix from Compose (aosp/1783329) for swipeable exception when offset is very close to an anchor.

Test: Run androidx.wear.compose.integration.demos.test
Bug: 222506652
Relnote: "SwipeToDismissBoxState has been refactored to restrict the scope of ExperimentalWearMaterialApi to Modifier.swipeable and SwipeableState, which are now used internally. SwipeToDismissBoxState now has currentValue, targetValue, isAnimationRunning and snapTo members to support common use cases - please let us know if you require any further properties to be made available. Also fixed the behavior of SwipeableState in the case where the swipe offset is within a rounding error of an anchor."
Change-Id: I58302544971da7b904a0b5681c06aae8e3f2ce3f