Fix StudioTask for playground
The playground version of the studio task was not finding the right
paths for idea vm options and properties. On mac, this fails to start
studio (haven't tried linux).
It was broken in I6d4c24e611469627f1e3c6c869e09b7548a8e647.

This CL fixes it to build the correct path.
I've also updated StudioTask to check if the files exists so that we can
find such problems easier. Furthermore, I've updated it to use
canonicalPath instead of absolutePath for clarity.

Bug: n/a
Test: ./studio m and also cd collections && ./gradlew studio

Change-Id: Icfc145958640c3fcadf04a50f903a5c0b19ecb40