Base visible entries off max lifecycle
Instead of having visible entries based off of the entry current
lifecycle which considers the host lifecycle, we should base them on the
maxLifecycle.

This means in the case of nested NavHost, the visible entries of the
inner NavHost will remain the same even if the outer state goes below
STARTED.

RelNote: "The `visibleEntries` `StateFlow provided by `NavController` is now based on the entry max Lifecycle state instead of the current Lifecycle state. This means that even if the host lifecycle of the navController goes below STARTED, the list of visible entires will remain the same."
Test: added test
Bug: 225394514

Change-Id: I9e2a810f261e9dc1e61417b168427b3bf597fbd5