Use visibleEntries in NavHost
Instead of all the custom code we currently use in NavHost to determine
which destinations to compose, we can just use visibleEntries directly
which has already sorted those destinations for us.

Moved the custom APIs over to DialogHost to fix that separately as the
DialogHost does not have access to a NavController.

RelNote: "NavHost now depend on `visibleEntries` from the NavController
to determine which entries to compose. This means that when using nested
NavHost the inner NavHost should now properly animate out."
Test: all existing tests pass
Bug: 225394514

Change-Id: I4ba2b542821154a76566f9c02cf7a95599a30f14