Introduce CanvasRenderer2 and GlesRenderer2
We now pass non-nullable sharedAssets using the user's type to
render and there's no longer a property which gets set
asynchronously. This fixes all the initial API concerns.

Bug: 223374561
Test: Manual testing + new tests
Relnote: Renderer.CanvasRenderer and Renderer.GlesRenderer have been deprecated in favor of Renderer.CanvasRenderer2 and Renderer.GlesRenderer2 which support SharedAssets which are passed to the render methods. For java interop we've introduced ListenableCanvasRenderer2 and ListenableGlesRenderer2.
Change-Id: I31ffac1033664bdb1c85eea5389ddb8136d1b59e