Introduce Snapshot.withoutReadObservation
Relnote: New function Snapshot.withoutReadObservation { ... } was added. It allows to run the passed lambda without subscribing to the changes of the state values read during this block. You could find it useful in use cases when you want to benefit from the snapshot based thread safe write/reads, but want to be able to read the value without causing unnecessary recomposition or remeasure.
Fixes: 214054486
Test: new tests in SnapshotStateObserverTestsJvm and RecomposerTests, rerun tests in ui module
Change-Id: I9f365d653483310cfda02cfa2c582fdcce8cfe33