Enable ProcessingCaptureSession to set/get tag and get capture result
When ProcessingCaptureSession is used (advanced extensions enabled),
Do the following change upon setRepeating
(1) set tags from SessionConfig, able to get tags in the capture
    result
(2) ensure CameraCaptureCallback in SessionConfig can be called and
   can use Camera2 CaptureCallback to receive the CaptureRequest,
   CaptureResult parameters.

After the change, setRepeating call in the ProcessingCaptureSession
will be able to invoke the CameraCaptureCallback with camera2
CaptureResult and CaptureRequest(for tags) parameters just like the normal
CaptureSession.

This will ensure zoom / focusMetering / flash work properly when
advanced extension is enabled.

But still capture still has the restriction that it cannot get
CaptureResult and CaptureRequest in the callback.

Test: ProcessingCaptureSessionTest
Bug: 183683118
Change-Id: I26149489b12f61724ba29b5e3ad4ec776e0ec67a