ToggleChip/SplitToggle chip: Remove default icon creation to ensure that developers set contentDescription
In order to ensure that developers are considering i18n and a11y we need to ensure that developers are required to set contentDescription on non-decorative icons.

This change removes the defaulting of the toggle control in ToggleChip and SplitToggleChip and also changes the Switch/Radio/CheckboxIcon methods that return Icons to instead return ImageVectors which the developer can then use to create their own icon. This is consistent with other changes we recently made to Slider/Stepper in order to ensure that developers 1) have a way to set contentDescription, 2) are required to set it and 3) use the standard consistent core compose means to create the icons rather than having a different helper API to do it.

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 228018150
RelNote: "In order to better support i18n and a11y we have changed ToggleChip and SplitToggleChip so that they no longer have a default for the toggleControl slot. We have also changed ToggleChipDefaults so that the following methods now return ImageVector rather than Icon (note that as they no longer return @Composables they have changed to start with lower case), SwitchIcon()->switchIcon(), CheckboxIcon->checkboxIcon() and RadioIcon()->radioIcon() - this allows and encourages developers to create their own Icon() composables with an appropriate contentDescription set."
Change-Id: I5bb5b4158b2818130c314e4020249d96d4da59fb