Fix bug for a CurvedLayout changing children on a SideEffect.
Previously, the internal "curved tree" was updated in the main
dereivedStateOf on CurvedLayout when a container changed it's content.
This may cause it to be out of sync with compose's tree.
Now, we update our tree during composition, so they are kept sync.
Relnote: "Bug fix for https://issuetracker.google.com/issues/226648931"
Test: Added regresion test
Bug: 226648931

Change-Id: Ia0a0aa4e2e480814665f0687af41506842f36de6