Move SavedState, Lifecycle to Beta01
This is a follow up to ag/17599650.  It only actually included the
API files but didn't actually increment the version.  We don't want
androidx.activity to bump to beta, but we do need it for SavedState,
and Lifecycle in order to be able to release androidx.activity
from androidx-platform-dev.

Change-Id: I119867bad1508dccf05ea0f69da0760c152b71a0
RelNote: N/A
Test: ./gradlew checkApi