Add Brush to TextStyle and SpanStyle.
Brush is a more capable way of drawing shapes, geometry, and text in Compose, a superset of Color. This change brings Brush as a first class property to TextStyle and SpanStyle to enable drawing Text composables using Brush.

Fix: b/187839528
Relnote: Added Brush to TextStyle and SpanStyle to provide a way to draw text with gradient coloring.
Test: Manual tests on Pixel 4a/6 Pro
Test: New tests
Test: ui:ui-text:cAT
Change-Id: I538698c22f5a03b57ed811ea733dd1194deaaa6a