WV: Fix force dark tests failure on Android T.
It seems that the default Android T theme darkens the view
background color, it cause "about:blank" which has transparent
background become dark.

Instead of relying on platform's default theme, this patch
changes it to use WebViewLightThemeTestActivity.

This patch also changes the expected image histogram to match it
to the new activity theme.

Bug: 229266593
Test: Pass the existing tests
Change-Id: Ibf2df9d6fdcce18c3578b25cf73cb41d1699dbfd