Fixes Broken AnnotationProcessorLocalTest
When combining annotations, processing of some @Document classes needs
to be deferred until the corresponding @AutoValue class is generated.
Previously, the AppSearchCompiler would only be able to process one
class with both annotations. Now, multiple classes may be deferred.

On some machines, this fix had no effect as AnnotationProcessorLocalTest
always ran fine. To be sure, we had to test this on a machine where
AnnotationProcessorLocalTest was failing before we made these changes to
AppSearchCompiler.

Test: Run AppSearchCompilerTest, AnnotationProcessorLocalTest
Bug: 229301043
Bug: 176202881
Change-Id: I258d1eecf59a8a763edeec7c49e90e7c3c08a76c