Fix crash with nested SubcomposeLayouts in some conditions
After the recent change aosp/2056467 SubcomposeLayout`s children left for reuse are considered disposed. If there were nested SubcomposeLayout in this reusable slot it ended up in the inconsistent condition where the inner state things there are no slots composed, but in fact we have attached children. In simple use cases it will not be surfaced, but if this SubcomposeLayout will try to remeasure it will detect the inconsistent state and crash. This change makes sure the state is consistent - we are not only clearing the inner state, but also removing the layout children.

Fixes: 229752147
Test: new tests in SubcomposeLayoutTest
Change-Id: I7f591f73696aa88c704050d4cf8298cde12d4254