Cleanup JavaCompileInputs's KMP path
This CL updates JavaCompileInputs to build the list of source
folders based on KotlinPlatformType instead of hardcoding
sourceSet names. These sourceSet names can be arbitrary hence
their names are not a great way to discover jvm/android source
sets.

I've also noticed the ApiTasks was doing a very similar and hardcoded
logic so I moved it inside the JavaCompileInputs as well.

Bug: n/a
Test: CI
Change-Id: Id38dfbf9ad5b62560a2271f432f3d4f4062f31e3